% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show accident density in a given state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numerical value corresponding to a state in the FARS data frame}
\item{year}{A numeric value denoting the year during which data should be gathered}
}
\value{
This function plots the location of the accidents on a map of a given state in a given year
}
\description{
This is a function that generates a data frame for a given year (\code{years}) using the functions \code{\link{make_filename}} and \code{\link{fars_read}}
and plots the accidents for a given state (\code{state.num}).
An error message is printed if \code{state.num} does not appear in the FARS data set ("invalid STATE number")
or if no accidents appeared ("no accidents to plot")
}
