% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads in data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string representing the name of the file to read in}
}
\value{
This function returns a data frame from the file input.
}
\description{
This is a function that reads in a FARS data file (with file name \code{filename}).
An error message is generated if the argument \code{filename} does not exist.
}
