% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Generate data frame and select months and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric value denoting the year/ a list of years during which data should be gathered}
}
\value{
This function returns a data frame containing month and year for each data entry of the corresponding data set.
}
\description{
This is a function that generates a data frame for given \code{years} (from the FARS data file)
using the functions \code{\link{make_filename}} and \code{link{fars_read}} and selects only the columns for month and year.
An error message ("invalid year") is printed if an associated data set cannot be found.
}
