% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Yields sum of accidents for each month in a given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric value denoting the year/ a list of years during which data should be gathered}
}
\value{
This function returns a data frame containing the number of accidents for each month in a given year.
}
\description{
This is a function that summarizes the data frame generated by the function \code{\link{fars_read_years}}
yielding the sum of accidents for each month for given \code{years}.
}
