% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-functions.R
\name{make_filename}
\alias{make_filename}
\title{Generates file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value denoting the year during which data should be gathered}
}
\value{
This function returns a character string (the file name) corresponding to the given \code{year}.
}
\description{
This is a function that generates for a given \code{year} the appropriate file name for the corresponding FARS data file.
}
